api, ocr shim: one rule for labeled creates; a finish EDT releases its own token
The public hint loses its check field: a create of a live GUID parks,
whatever properties it carries, and no creator is told anything. The
hint's finish-scope documentation now states the contract the runtime
enforces in Debug builds -- the scope must be homed on the creating
rank -- and a caller that owns a finish scope's creator token can
release it early.

The shim maps every labeled create (data block, event, EDT) onto that
one rule: GUID_PROP_CHECK and GUID_PROP_BLOCK are accepted and ignored,
ocrEdtCreate honours a labeled EDT GUID, and a collective's edges are
created by the consuming generation only, with producers satisfying
what the consumer created. An OCR finish EDT's output event must fire
when the finish EDT and its descendants complete, so the shim releases
the finish EDT's creator token right after the finish EDT's own join
rather than at the end of the creating task; the shim's own test pins
that, in the shim's build.

(cherry picked from commit 23103e7775657ed058e7a0f4637f7c1ddfb5f7bf)

+++ tests/shim/event_payload.c
@@ -14,10 +14,10 @@ static ocrGuid_t check_payload(u32 pc, u64 *pv, u32 dc, ocrEdtDep_t dv[]) {
   int ok[7];
   ok[0] = ocrGuidIsNull(dv[0].guid) && !dv[0].ptr;
   ok[1] = ocrGuidIsNull(dv[1].guid) && !dv[1].ptr;
-  ok[2] = dv[2].guid.guid == pv[0] && dv[2].ptr && *(u64 *)dv[2].ptr == 42;
-  ok[3] = dv[3].guid.guid == pv[0] && dv[3].ptr && *(u64 *)dv[3].ptr == 42;
-  ok[4] = dv[4].guid.guid == pv[0] && !dv[4].ptr;
-  ok[5] = dv[5].guid.guid == pv[1] && !dv[5].ptr;
+  ok[2] = (u64)dv[2].guid.guid == pv[0] && dv[2].ptr && *(u64 *)dv[2].ptr == 42;
+  ok[3] = (u64)dv[3].guid.guid == pv[0] && dv[3].ptr && *(u64 *)dv[3].ptr == 42;
+  ok[4] = (u64)dv[4].guid.guid == pv[0] && !dv[4].ptr;
+  ok[5] = (u64)dv[5].guid.guid == pv[1] && !dv[5].ptr;
   ok[6] = ocrGuidIsNull(dv[6].guid) && !dv[6].ptr;
   for (u32 i = 0; i < 7 && i < dc; ++i) {
     if (!ok[i]) {
